shared find employee function
