+++ src/functions/shared.function.ts
@@ -10,3 +10,28 @@ export const getEmployee = async (employeeUuid: string) => {
 	});
 	return {status: 200, data: employee};
 };
+
+export const findEmployee = async (
+	companyUuid: string,
+	keyword: any, //TODO define keyword type
+) => {
+	const employees = await prisma.company.findMany({
+		where: {
+			uuid: companyUuid,
+		},
+		select: {
+			Employee: {
+				where: {
+					fullName: {
+						contains: keyword,
+					},
+				},
+				take: 5,
+				orderBy: {
+					createdAt: "desc",
+				},
+			},
+		},
+	});
+	return {status: 200, data: employees};
+};
